Added error handling, delete confirmation and update CSS.

+++ client/src/contexts/StateContext.js
@@ -10,6 +10,7 @@ export const StateProvider = ({
 }) => {
     const navigate = useNavigate();
     const [posts, setPosts] = useState([]);
+    const [error, SetError] = useState(null);
 
     useEffect(() => {
         postsService.getAllPosts()
@@ -19,34 +20,58 @@ export const StateProvider = ({
     }, []);
 
     const onCreatePostSubmit = async (data) => {
-        const post = await postsService.createPost(data);
-       
-        setPosts(state => [post, ...state]);
+        try {
+            if (data.title === '' || data.imageUrl === '' || data.author === '' || data.content === '') {
+                throw new Error('All fields are required!');
+            }
+            SetError('');
 
-        navigate('/catalog');
+            const post = await postsService.createPost(data);
+
+            setPosts(state => [post, ...state]);
+
+            navigate('/catalog');
+        } catch (err) {
+            SetError(err);
+        }
     };
 
     const onEditPostSubmit = async (values) => {
-        const result = await postsService.editPost(values._id, values);
+        try {
+            if (values.title === '' || values.imageUrl === '' || values.author === '' || values.content === '') {
+                throw new Error('All fields are required!');
+            }
+            SetError('');
 
-        setPosts(state => state.map(x => x._id === values._id ? result : x));
-        
-        navigate(`/catalog/${values._id}`);
+            const result = await postsService.editPost(values._id, values);
+
+            setPosts(state => state.map(x => x._id === values._id ? result : x));
+
+            navigate(`/catalog/${values._id}`);
+        } catch (err) {
+            SetError(err);
+        }
     };
 
     const onDelete = async (postId) => {
-        await postsService.deletePost(postId);
+        // eslint-disable-next-line no-restricted-globals
+        const choice = confirm('Are you sure you want to delete this post?');
 
-        setPosts(state => state.filter(x => x._id !== postId));
-      
-        navigate('/catalog');
+        if (choice) {
+            await postsService.deletePost(postId);
+
+            setPosts(state => state.filter(x => x._id !== postId));
+
+            navigate('/catalog');
+        }
     };
 
     const ctxState = {
         onCreatePostSubmit,
         onEditPostSubmit,
         onDelete,
-        posts
+        posts,
+        error
     };
 
     return (
